fix:리그 로직 수적

+++ src/main/java/io/api/myasset/domain/home/dto/response/HomeRankingPercentileResponse.java
@@ -9,6 +9,6 @@ import lombok.Getter;
 @Builder
 public class HomeRankingPercentileResponse {
 
-	private final Integer rankingPercentile;
+	private final double  rankingPercentile;
 	private final LocalDate measuredAt;
 }
